Moved data sheet deps around
