+++ frontend/package.json
@@ -15,7 +15,8 @@
     "@blueprintjs/select": "^3.15.5",
     "@blueprintjs/table": "^3.8.23",
     "@data-ui/histogram": "0.0.75",
-    "@earthcube/schema-linker": "*",
+    "@earthdata/schema-linker": "*",
+    "@earthdata/sheet": "*",
     "@emotion/core": "^10.0.7",
     "@emotion/styled": "^10.0.7",
     "@loadable/component": "^5.10.2",
