+++ frontend/src/admin/data-sheet/index.tsx
@@ -4,12 +4,12 @@ import { useAPIResult } from "@macrostrat/ui-components";
 import update from "immutability-helper";
 import h from "@macrostrat/hyper";
 import { Frame } from "~/frame";
-import { DataSheetProvider } from "./provider";
+import { DataSheetProvider } from "@earthdata/sheet/src/provider";
 import { SheetToolbar } from "./toolbar";
 import { VirtualizedSheet } from "./virtualized";
 import classNames from "classnames";
 import styles from "./module.styl";
-import { Row, Sheet } from "./components";
+import { Row, Sheet } from "@earthdata/sheet/src/components";
 import { useElementSize } from "./util";
 import { sum } from "d3-array";
 import { combineLikeIds, addNecesaryFields } from "./util";
