+++ frontend/src/admin/data-sheet/virtualized.ts
@@ -2,7 +2,7 @@ import { hyperStyled } from "@macrostrat/hyper";
 import { useContext, useRef, useState } from "react";
 import { useElementSize, useScrollOffset } from "./util";
 import ReactDataSheet from "react-datasheet";
-import { DataSheetContext, DataSheetProvider } from "./provider";
+import { DataSheetContext } from "@earthdata/sheet/src/provider";
 import styles from "./module.styl";
 
 const h = hyperStyled(styles);
