+++ frontend/tsconfig.json
@@ -16,5 +16,12 @@
       "~/*": ["src/*"]
     }
   },
-  "include": ["src", "plugins", "default-content"]
+  "include": [
+    "src",
+    "plugins",
+    "default-content",
+    "packages/sheet/src/components.ts",
+    "packages/sheet/src/virtualized.ts",
+    "packages/sheet/src/provider.ts"
+  ]
 }
